third (and final) try to get a reasonable speed up

+++ labs/memory_bound/data_packing/init.cpp
@@ -5,10 +5,10 @@
 S create_entry(int first_value, int second_value) {
   S entry;
 
-  entry.i = first_value;
-  entry.s = static_cast<short>(second_value);
-  entry.l = static_cast<long long>(first_value * second_value);
-  entry.d = static_cast<double>(first_value) / maxRandom;
+  entry.i = static_cast<uint8_t>(first_value);
+  entry.s = static_cast<uint8_t>(second_value);
+  entry.l = static_cast<uint16_t>(first_value * second_value);
+  entry.d = static_cast<float>(first_value) / maxRandom;
   entry.b = first_value < second_value;
 
   return entry;
